Restructure zigbee-lock
First, I had AI do the following:
- Rename subdrivers: using-old-capabilities -> legacy-handlers,
  promote using-new-capabilities handlers to top-level defaults
- Make new capability handling the driver default; legacy-handlers
  subdriver overrides for pre-migration devices only
- Top-level vendor subdrivers (yale, samsungsds, yale-fingerprint-lock)
  now only activate for migrated devices to avoid double-dispatch
- Remove duplicate lock-without-codes top-level subdriver (handled
  entirely within legacy-handlers since these devices never migrate)
- Add init lifecycle handler to legacy-handlers to populate lockCodes
  state for pre-migration devices on driver restart


Then, I cleaned up the driver further by fixing the git understanding of
what files were made in what year, and by introducing small changes to
altogether remove some sub-subdrivers from the legacy subdriver, and vice versa.
