+++ drivers/SmartThings/zigbee-lock/src/bad-battery-reporter/init.lua
@@ -1,0 +1,26 @@
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+
+local clusters = require "st.zigbee.zcl.clusters"
+local capabilities = require "st.capabilities"
+
+
+
+local battery_report_handler = function(driver, device, value)
+   device:emit_event(capabilities.battery.battery(value.value))
+end
+
+local bad_battery_reporter_driver = {
+  NAME = "Bad Battery Reporter Driver",
+  zigbee_handlers = {
+    attr = {
+      [clusters.PowerConfiguration.ID] = {
+        [clusters.PowerConfiguration.attributes.BatteryPercentageRemaining.ID] = battery_report_handler
+      }
+    }
+  },
+  can_handle = require("bad-battery-reporter.can_handle")
+}
+
+return bad_battery_reporter_driver
