+++ drivers/SmartThings/zigbee-lock/src/legacy-handlers/can_handle.lua
@@ -1,0 +1,13 @@
+-- Copyright 2026 SmartThings
+-- Licensed under the Apache License, Version 2.0
+
+return function(opts, driver, device, ...)
+  local capabilities = require "st.capabilities"
+  local lock_codes_migrated = device:get_latest_state("main", capabilities.lockCodes.ID,
+    capabilities.lockCodes.migrated.NAME, false)
+  if not lock_codes_migrated then
+    local subdriver = require("legacy-handlers")
+    return true, subdriver
+  end
+  return false
+end
