+++ drivers/SmartThings/zigbee-lock/src/legacy-handlers/lock-without-codes/fingerprints.lua
