+++ drivers/SmartThings/zigbee-lock/src/legacy-handlers/yale/can_handle.lua
@@ -1,0 +1,11 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local function yale_can_handle(opts, driver, device, ...)
+  if device:get_manufacturer() == "ASSA ABLOY iRevo" or device:get_manufacturer() == "Yale" then
+    return true, require("legacy-handlers.yale")
+  end
+  return false
+end
+
+return yale_can_handle
