+++ drivers/SmartThings/zigbee-lock/src/samsungsds/init.lua
@@ -1,0 +1,118 @@
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+
+local device_management = require "st.zigbee.device_management"
+local clusters = require "st.zigbee.zcl.clusters"
+local battery_defaults = require "st.zigbee.defaults.battery_defaults"
+local capabilities = require "st.capabilities"
+local cluster_base = require "st.zigbee.cluster_base"
+local PowerConfiguration = clusters.PowerConfiguration
+local DoorLock = clusters.DoorLock
+local Lock = capabilities.lock
+local lock_utils = require "zigbee_lock_utils"
+
+local SAMSUNG_SDS_MFR_SPECIFIC_UNLOCK_COMMAND = 0x1F
+local SAMSUNG_SDS_MFR_CODE = 0x0003
+
+local function handle_lock_state(driver, device, value, zb_rx)
+  if value.value == DoorLock.attributes.LockState.LOCKED then
+    device:emit_event(Lock.lock.locked())
+  elseif value.value == DoorLock.attributes.LockState.UNLOCKED then
+    device:emit_event(Lock.lock.unlocked())
+  end
+end
+
+local function mfg_lock_door_handler(driver, device, zb_rx)
+  local cmd = zb_rx.body.zcl_body.body_bytes:byte(1)
+  if cmd == 0x00 then
+    device:emit_event(Lock.lock.unlocked())
+  end
+end
+
+local function unlock_cmd_handler(driver, device, command)
+  device:send(cluster_base.build_manufacturer_specific_command(
+          device,
+          DoorLock.ID,
+          SAMSUNG_SDS_MFR_SPECIFIC_UNLOCK_COMMAND,
+          SAMSUNG_SDS_MFR_CODE,
+          "\x10\x04\x31\x32\x33\x35"))
+end
+
+local function lock_cmd_handler(driver, device, command)
+  -- do nothing in lock command handler
+end
+
+local refresh = function(driver, device, cmd)
+  -- do nothing in refresh capability handler
+end
+
+local function emit_event_if_latest_state_missing(device, component, capability, attribute_name, value)
+  if device:get_latest_state(component, capability.ID, attribute_name) == nil then
+    device:emit_event(value)
+  end
+end
+
+local function load_device_state(device)
+  local lock_codes_migrated = device:get_latest_state("main", capabilities.lockCodes.ID, capabilities.lockCodes.migrated.NAME, false)
+  if lock_codes_migrated then
+    lock_utils.reload_tables(device)
+  else
+    local legacy_lock_utils = require "legacy-handlers.legacy_lock_utils"
+    legacy_lock_utils.populate_state_from_data(device)
+  end
+end
+
+local device_added = function(self, device)
+  load_device_state(device)
+  emit_event_if_latest_state_missing(device, "main", capabilities.lock, capabilities.lock.lock.NAME, capabilities.lock.lock.unlocked())
+  device:emit_event(capabilities.battery.battery(100))
+end
+
+local do_configure = function(self, device)
+  device:send(device_management.build_bind_request(device, DoorLock.ID, self.environment_info.hub_zigbee_eui))
+  device:send(device_management.build_bind_request(device, PowerConfiguration.ID, self.environment_info.hub_zigbee_eui))
+  device:send(DoorLock.attributes.LockState:configure_reporting(device, 0, 3600, 0))
+end
+
+local battery_init = battery_defaults.build_linear_voltage_init(4.0, 6.0)
+
+local device_init = function(driver, device, event)
+  battery_init(driver, device, event)
+  device:remove_monitored_attribute(clusters.PowerConfiguration.ID, clusters.PowerConfiguration.attributes.BatteryVoltage.ID)
+  device:remove_configured_attribute(clusters.PowerConfiguration.ID, clusters.PowerConfiguration.attributes.BatteryVoltage.ID)
+  load_device_state(device)
+end
+
+local samsung_sds_driver = {
+  NAME = "SAMSUNG SDS Lock Driver",
+  zigbee_handlers = {
+    cluster = {
+      [DoorLock.ID] = {
+        [SAMSUNG_SDS_MFR_SPECIFIC_UNLOCK_COMMAND] = mfg_lock_door_handler
+      }
+    },
+    attr = {
+      [DoorLock.ID] = {
+        [DoorLock.attributes.LockState.ID] = handle_lock_state
+      }
+    }
+  },
+  capability_handlers = {
+    [capabilities.refresh.ID] = {
+      [capabilities.refresh.commands.refresh.NAME] = refresh
+    },
+    [capabilities.lock.ID] = {
+      [capabilities.lock.commands.unlock.NAME] = unlock_cmd_handler,
+      [capabilities.lock.commands.lock.NAME] = lock_cmd_handler
+    }
+  },
+  lifecycle_handlers = {
+    doConfigure = do_configure,
+    added = device_added,
+    init = device_init
+  },
+  can_handle = require("samsungsds.can_handle")
+}
+
+return samsung_sds_driver
