+++ drivers/SmartThings/zigbee-lock/src/sub_drivers.lua
@@ -3,7 +3,9 @@
 
 local lazy_load_if_possible = require "lazy_load_subdriver"
 local sub_drivers = {
-   lazy_load_if_possible("using-old-capabilities"),
-   lazy_load_if_possible("using-new-capabilities"),
+   lazy_load_if_possible("legacy-handlers"),
+   lazy_load_if_possible("samsungsds"),
+   lazy_load_if_possible("bad-battery-reporter"),
+   lazy_load_if_possible("yale-fingerprint-lock"),
 }
 return sub_drivers
