+++ drivers/SmartThings/zigbee-lock/src/yale-fingerprint-lock/fingerprints.lua
